requirements.txt
psycopg2-binary>=2.9
pandas>=2.0
numpy>=1.26
faker>=25.0